refactor(tests): reuse transit agency fixture
