--- tests/e2e/cypress/fixtures/transit-agencies.json
@@ -1,20 +1,0 @@
-[
-  {
-    "slug": "abc",
-    "short_name": "ABC",
-    "long_name": "ABC Transit Company",
-    "agency_id": "abc123",
-    "merchant_id": "abc",
-    "info_url": "https://www.example.com/help",
-    "phone": "800-555-5555"
-  },
-  {
-    "slug": "deftl",
-    "short_name": "DefTL",
-    "long_name": "DEF Transit Lines",
-    "agency_id": "def456",
-    "merchant_id": "deftl",
-    "info_url": "https://www.example.com/help",
-    "phone": "321-555-5555"
-  }
-]
